+++ Text/MMark.hs
@@ -8,12 +8,12 @@
 -- Portability :  portable
 --
 -- MMark (read “em-mark”) is a strict markdown processor for writers.
--- “Strict” means that not every input is considered a valid markdown
--- document and parse errors are possible and even desirable, because they
--- allow to spot markup issues without searching for them in rendered
--- document. If a markdown document passes MMark parser, then it'll most
--- certainly produce HTML without quirks. This feature makes it a good
--- choice for writers and bloggers.
+-- “Strict” means that not every input is considered valid markdown document
+-- and parse errors are possible and even desirable, because they allow to
+-- spot markup issues without searching for them in rendered document. If a
+-- markdown document passes MMark parser, then it'll likely produce HTML
+-- without quirks. This feature makes it a good choice for writers and
+-- bloggers.
 --
 -- === MMark and Common Mark
 --
@@ -22,9 +22,9 @@
 -- <https://github.com/jgm/CommonMark>
 --
 -- However, due to the fact that we do not allow inputs that do not make
--- sense, and also try to guard against common silly mistakes (like writing
--- @##My header@ and having it rendered as a paragraph starting with hashes)
--- MMark obviously can't follow the specification precisely. In particular,
+-- sense, and also try to guard against common mistakes (like writing @##My
+-- header@ and having it rendered as a paragraph starting with hashes) MMark
+-- obviously can't follow the specification precisely. In particular,
 -- parsing of inlines differs considerably from Common Mark.
 --
 -- Another difference between Common Mark and MMark is that the latter